Move extension methods to util
