+++ app/src/main/java/com/github/pockethub/android/ui/item/gist/GistItem.kt
@@ -4,7 +4,7 @@ import android.text.TextUtils
 import androidx.text.bold
 import androidx.text.buildSpannedString
 import com.github.pockethub.android.R
-import com.github.pockethub.android.android.text.append
+import com.github.pockethub.android.util.android.text.append
 import com.github.pockethub.android.util.AvatarLoader
 import com.meisolsson.githubsdk.model.Gist
 import com.xwray.groupie.kotlinandroidextensions.Item
