+++ app/src/main/java/com/github/pockethub/android/ui/item/news/PushEventItem.kt
@@ -4,7 +4,7 @@ import android.text.TextUtils
 import android.view.View
 import androidx.text.bold
 import androidx.text.buildSpannedString
-import com.github.pockethub.android.android.text.monospace
+import com.github.pockethub.android.util.android.text.monospace
 import com.github.pockethub.android.ui.view.OcticonTextView
 import com.github.pockethub.android.util.AvatarLoader
 import com.meisolsson.githubsdk.model.GitHubEvent
